+++ record_platform_interface/lib/src/types/record_state.dart
@@ -1,0 +1,2 @@
+/// All states of the recorder
+enum RecordState { pause, record, stop }