+++ record_platform_interface/lib/src/types/types.dart
@@ -1,3 +1,4 @@
 export 'package:record_platform_interface/src/types/amplitude.dart';
 export 'package:record_platform_interface/src/types/audio_encoder.dart';
 export 'package:record_platform_interface/src/types/input_device.dart';
+export 'package:record_platform_interface/src/types/record_state.dart';