+++ record_windows/lib/record_windows.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:math';
 
+import 'package:flutter/foundation.dart';
 import 'package:path/path.dart' as p;
 import 'package:record_platform_interface/record_platform_interface.dart';
 
@@ -21,6 +22,8 @@ class RecordWindows extends RecordPlatform {
   bool _isRecording = false;
   bool _isPaused = false;
   String? _path;
+  StreamSubscription<String>? _fmediaSub;
+  StreamController<RecordState>? _stateStreamCtrl;
 
   @override
   Future<void> dispose() {
@@ -29,6 +32,8 @@ class RecordWindows extends RecordPlatform {
         Process.killPid(_pid!, ProcessSignal.sigterm);
         _pid = null;
       }
+      _fmediaSub?.cancel();
+      _stateStreamCtrl?.close();
       return Future.value();
     });
   }
@@ -47,13 +52,10 @@ class RecordWindows extends RecordPlatform {
   Future<bool> isEncoderSupported(AudioEncoder encoder) async {
     switch (encoder) {
       case AudioEncoder.aacLc:
-        return true;
+      case AudioEncoder.aacHe:
       case AudioEncoder.flac:
-        return true;
       case AudioEncoder.opus:
-        return true;
       case AudioEncoder.wav:
-        return true;
       case AudioEncoder.vorbisOgg:
         return true;
       default:
@@ -75,14 +77,14 @@ class RecordWindows extends RecordPlatform {
   Future<void> pause() async {
     await _callFMedia(['--globcmd=pause']);
 
-    _isPaused = true;
+    _updateState(RecordState.pause);
   }
 
   @override
   Future<void> resume() async {
     await _callFMedia(['--globcmd=unpause']);
 
-    _isPaused = false;
+    _updateState(RecordState.record);
   }
 
   @override
@@ -110,6 +112,7 @@ class RecordWindows extends RecordPlatform {
     _path = path;
 
     final process = await _callFMedia([
+      '--notui',
       '--background',
       '--record',
       '--out=$path',
@@ -122,7 +125,24 @@ class RecordWindows extends RecordPlatform {
     ]);
     _pid = process.pid;
 
-    _isRecording = true;
+    _updateState(RecordState.record);
+
+    // Print the fmedia output to allow diagnostics.
+    if (kDebugMode) {
+      _fmediaSub?.cancel();
+      _fmediaSub = process.stderr
+          .transform(utf8.decoder)
+          .transform(const LineSplitter())
+          .listen((chunk) {
+        if (kDebugMode) {
+          print(chunk);
+        }
+      })
+        ..onDone(() async {
+          _fmediaSub?.cancel();
+          stop();
+        });
+    }
   }
 
   @override
@@ -130,8 +150,7 @@ class RecordWindows extends RecordPlatform {
     await _callFMedia(['--globcmd=stop']);
     await _callFMedia(['--globcmd=quit']);
 
-    _isRecording = false;
-    _isPaused = false;
+    _updateState(RecordState.stop);
 
     return _path;
   }
@@ -156,6 +175,18 @@ class RecordWindows extends RecordPlatform {
       });
 
     return completer.future;
+  }
+
+  @override
+  Stream<RecordState> onStateChanged() {
+    _stateStreamCtrl ??= StreamController(
+      onCancel: () {
+        _stateStreamCtrl?.close();
+        _stateStreamCtrl = null;
+      },
+    );
+
+    return _stateStreamCtrl!.stream;
   }
 
   String _getFileNameSuffix(AudioEncoder encoder) {
@@ -183,7 +214,7 @@ class RecordWindows extends RecordPlatform {
       case AudioEncoder.aacHe:
         return ['--aac-profile=HEv2', ..._getAacQuality(bitRate)];
       case AudioEncoder.flac:
-        return ['--flac-compression=6'];
+        return ['--flac-compression=6', '--format=int16'];
       case AudioEncoder.opus:
         final rate = (bitRate ~/ 1000).clamp(6, 510);
         return ['--opus.bitrate=$rate'];
@@ -298,5 +329,26 @@ class RecordWindows extends RecordPlatform {
       channels: channels,
       samplingRate: samplingRate,
     );
+  }
+
+  Future<void> _updateState(RecordState state) async {
+    switch (state) {
+      case RecordState.pause:
+        _isRecording = true;
+        _isPaused = true;
+        break;
+      case RecordState.record:
+        _isRecording = true;
+        _isPaused = false;
+        break;
+      case RecordState.stop:
+        _isRecording = false;
+        _isPaused = false;
+        break;
+    }
+
+    if (_stateStreamCtrl?.hasListener ?? false) {
+      _stateStreamCtrl?.add(state);
+    }
   }
 }