+++ record_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.5.0
+- feat: onStateChanged is now provided by platform implementations.
+
 ## 0.4.0
 - feat: Add input devices listing.
 - feat: Add number of channels.